requirements.txt
streamlit
torch
torchvision
numpy
matplotlib
pillow
scikit-image
coremltools
kagglehub
opencv-python
